app/requirements.txt
Django==3.1.1
pytube3==9.6.4
ffmpeg==1.4
gunicorn==20.0.4